reddit_kafka_producer/requirements.txt
kafka-python==2.0.2
praw==7.6.0
python-dotenv==0.19.2